+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -14,26 +14,8 @@ public enum FeatureFlag: String, CaseIterable {
     /// Whether End Of Year feature is enabled
     case endOfYear
 
-    /// Enable show notes using the new endpoint
-    case newShowNotesEndpoint
-
-    /// Enable retrieving episode artwork from the RSS feed
-    case episodeFeedArtwork
-
-    /// Enable chapters to be loaded from the RSS feed
-    case rssChapters
-
-    /// Enable a quicker and more responsive player transition
-    case newPlayerTransition
-
     /// Avoid logging out user on non-authorization HTTP errors
     case errorLogoutHandling
-
-    /// Enable the ability to rate podcasts
-    case giveRatings
-
-    /// Enable selecting/deselecting episode chapters
-    case deselectChapters
 
     /// Store settings as JSON in User Defaults (global) or SQLite (podcast)
     case newSettingsStorage
@@ -49,11 +31,6 @@ public enum FeatureFlag: String, CaseIterable {
 
     /// Enable the AVExportSession parallel download of any playing episode
     case streamAndCachePlayingEpisode
-
-    case categoriesRedesign
-
-    /// show UpNext tab on the main tab bar
-    case upNextOnTabBar
 
     /// When enabled it updates the code on filter callback to use a safer method to convert unmanaged player references
     /// This is to fix this: https://a8c.sentry.io/share/issue/39a6d2958b674ec3b7a4d9248b4b5ffa/
@@ -73,11 +50,6 @@ public enum FeatureFlag: String, CaseIterable {
 
     /// Use the Accelerate framework to speed up custom effects
     case accelerateEffects
-
-    case newSharing
-
-    /// Enable the transcripts feature on podcasts episodes
-    case transcripts
 
     /// Enables the Kids banner
     case kidsProfile
@@ -155,6 +127,14 @@ public enum FeatureFlag: String, CaseIterable {
     /// Uses the episode IDs from the server's response rather than our local database IDs
     case useSyncResponseEpisodeIDs
 
+    ///Use html description for podcast details
+    case usePodcastHTMLDescription
+
+    /// Disables logout / keychain clearing when errors occur in the background
+    case avoidLogoutInBackground
+
+    case disablePrivateFeedSharing
+
     public var enabled: Bool {
         if let overriddenValue = FeatureFlagOverrideStore().overriddenValue(for: self) {
             return overriddenValue
@@ -173,19 +153,7 @@ public enum FeatureFlag: String, CaseIterable {
             false
         case .endOfYear:
             false
-        case .newShowNotesEndpoint:
-            false
-        case .episodeFeedArtwork:
-            false
-        case .rssChapters:
-            false
-        case .newPlayerTransition:
-            true
         case .errorLogoutHandling:
-            false
-        case .giveRatings:
-            false
-        case .deselectChapters:
             false
         case .newSettingsStorage:
             shouldEnableSyncedSettings
@@ -197,11 +165,7 @@ public enum FeatureFlag: String, CaseIterable {
             false
         case .streamAndCachePlayingEpisode:
             true
-        case .categoriesRedesign:
-            true
         case .defaultPlayerFilterCallbackFix:
-            true
-        case .upNextOnTabBar:
             true
         case .downloadFixes:
             true
@@ -210,10 +174,6 @@ public enum FeatureFlag: String, CaseIterable {
         case .whenPlayingOnlyUpdateEpisodeIfPlaybackFails:
             true
         case .accelerateEffects:
-            true
-        case .newSharing:
-            true
-        case .transcripts:
             true
         case .kidsProfile:
             false
@@ -259,6 +219,12 @@ public enum FeatureFlag: String, CaseIterable {
 			true
         case .useSyncResponseEpisodeIDs:
             true
+        case .usePodcastHTMLDescription:
+            true
+        case .avoidLogoutInBackground:
+            true
+        case .disablePrivateFeedSharing:
+            true
         }
     }
 
@@ -270,26 +236,14 @@ public enum FeatureFlag: String, CaseIterable {
     /// This should match a Firebase Remote Config Parameter name (key)
     public var remoteKey: String? {
         switch self {
-        case .deselectChapters:
-            "deselect_chapters_enabled"
         case .newAccountUpgradePromptFlow:
             "new_account_upgrade_prompt_flow"
         case .newSettingsStorage:
             shouldEnableSyncedSettings ? "new_settings_storage" : nil
         case .settingsSync:
             shouldEnableSyncedSettings ? "settings_sync" : nil
-        case .newShowNotesEndpoint:
-             "new_show_notes"
-         case .episodeFeedArtwork:
-             "episode_artwork"
-         case .rssChapters:
-             "rss_chapters"
-        case .categoriesRedesign:
-            "categories_redesign"
         case .defaultPlayerFilterCallbackFix:
             "default_player_filter_callback_fix"
-        case .upNextOnTabBar:
-            "up_next_on_tab_bar"
         default:
             rawValue.lowerSnakeCased()
         }