+++ PocketCastsTests/Tests/Utilities/SettingsTests.swift
@@ -9,11 +9,22 @@ final class SettingsTests: XCTestCase {
     private let userDefaultsSuiteName = "PocketCasts-SettingsTests"
 
     private var overriddenFlags = [FeatureFlag: Bool]()
+    private let defaultPlayerActions: [PlayerAction] = [.addBookmark,
+                                                        .markPlayed,
+                                                        .transcript,
+                                                        .effects,
+                                                        .sleepTimer,
+                                                        .routePicker,
+                                                        .download,
+                                                        .shareEpisode,
+                                                        .goToPodcast,
+                                                        .starEpisode,
+                                                        .chromecast,
+                                                        .archive]
 
     override func setUp() {
         super.setUp()
         UserDefaults.standard.removePersistentDomain(forName: userDefaultsSuiteName)
-        try? override(flag: .transcripts, value: false)
     }
 
     private func override(flag: FeatureFlag, value: Bool) throws {
@@ -60,17 +71,7 @@ final class SettingsTests: XCTestCase {
         SettingsStore.appSettings.playerShelf = [.known(.markPlayed), .unknown(unknownString)]
         Settings.updatePlayerActions([.addBookmark, .markPlayed])
 
-        XCTAssertEqual([.addBookmark,
-                        .markPlayed,
-                        .effects,
-                        .sleepTimer,
-                        .routePicker,
-                        .download,
-                        .starEpisode,
-                        .shareEpisode,
-                        .goToPodcast,
-                        .chromecast,
-                        .archive], Settings.playerActions(), "Player actions should exclude unknown actions and include defaults")
+        XCTAssertEqual(defaultPlayerActions, Settings.playerActions(), "Player actions should exclude unknown actions and include defaults")
         XCTAssertEqual([.known(.addBookmark), .known(.markPlayed), .unknown(unknownString)], SettingsStore.appSettings.playerShelf, "Player shelf should include unknowns at end")
 
         try reset(flag: .newSettingsStorage)
@@ -82,17 +83,7 @@ final class SettingsTests: XCTestCase {
         Settings.updatePlayerActions(PlayerAction.defaultActions.filter { $0.isAvailable }) // Set defaults
         Settings.updatePlayerActions([.addBookmark, .markPlayed])
 
-        XCTAssertEqual([.addBookmark,
-                        .markPlayed,
-                        .effects,
-                        .sleepTimer,
-                        .routePicker,
-                        .download,
-                        .starEpisode,
-                        .shareEpisode,
-                        .goToPodcast,
-                        .chromecast,
-                        .archive], Settings.playerActions(), "Player actions should include changes from update")
+        XCTAssertEqual(defaultPlayerActions, Settings.playerActions(), "Player actions should include changes from update")
 
         try reset(flag: .newSettingsStorage)
     }
@@ -110,17 +101,7 @@ final class SettingsTests: XCTestCase {
         try setupSettingsStore()
         SettingsStore.appSettings.importUserDefaults()
 
-        XCTAssertEqual([.addBookmark,
-                        .markPlayed,
-                        .effects,
-                        .sleepTimer,
-                        .routePicker,
-                        .download,
-                        .starEpisode,
-                        .shareEpisode,
-                        .goToPodcast,
-                        .chromecast,
-                        .archive], Settings.playerActions(), "Player actions should include changes from update")
+        XCTAssertEqual(defaultPlayerActions, Settings.playerActions(), "Player actions should include changes from update")
 
         try reset(flag: .newSettingsStorage)
     }