+++ podcasts/AppDelegate.swift
@@ -296,19 +296,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             self?.updateEndOfYearRemoteValue()
             self?.updateRemoteFeatureFlags()
             ServerConfig.avoidLogoutOnError = FeatureFlag.errorLogoutHandling.enabled
+            ServerConfig.avoidLogoutInBackground = FeatureFlag.avoidLogoutInBackground.enabled
         }
     }
 
     func updateRemoteFeatureFlags() {
         guard BuildEnvironment.current != .debug else { return }
         do {
-            if FeatureFlag.newPlayerTransition.enabled != Settings.newPlayerTransition {
-                // If the player transition changes we dismiss the full screen player
-                // Otherwise this might lead to crashes or weird behavior
-                appDelegate()?.miniPlayer()?.closeFullScreenPlayer()
-                try FeatureFlagOverrideStore().override(FeatureFlag.newPlayerTransition, withValue: Settings.newPlayerTransition)
-            }
-
             if FeatureFlag.errorLogoutHandling.enabled != Settings.errorLogoutHandling {
                 ServerConfig.avoidLogoutOnError = FeatureFlag.errorLogoutHandling.enabled
                 try FeatureFlagOverrideStore().override(FeatureFlag.errorLogoutHandling, withValue: Settings.errorLogoutHandling)