+++ podcasts/PlaybackQueue.swift
@@ -228,9 +228,11 @@ class PlaybackQueue: NSObject {
         cacheTopEpisode()
         updateUpNextInfo()
 
+        #if !APPCLIP
         if checkForAutoDownload {
             checkAllForAutoDownload()
         }
+        #endif
     }
 
     func nowPlayingEpisodeChanged() {
@@ -311,11 +313,12 @@ class PlaybackQueue: NSObject {
     // MARK: - Private Helpers
 
     func updateUpNextInfo() {
-        #if !os(watchOS)
+        #if !os(watchOS) && !APPCLIP
             WidgetHelper.shared.updateSharedUpNext()
         #endif
     }
 
+    #if !APPCLIP
     private func checkAllForAutoDownload() {
         if !Settings.downloadUpNextEpisodes() { return }
 
@@ -337,6 +340,7 @@ class PlaybackQueue: NSObject {
             DownloadManager.shared.queueForLaterDownload(episodeUuid: episode.uuid, fireNotification: true, autoDownloadStatus: .autoDownloaded)
         }
     }
+    #endif
 
     private func cacheTopEpisode() {
         topEpisode = episodeAt(index: -1)