+++ podcasts/PlayerContainerViewController+Gestures.swift
@@ -7,51 +7,6 @@ extension PlayerContainerViewController: UIGestureRecognizerDelegate {
     private static let minimumScreenRatioToHide: CGFloat = 0.1
 
     @IBAction func panGestureRecognizerHandler(_ sender: UIPanGestureRecognizer) {
-        FeatureFlag.newPlayerTransition.enabled ? newTransitionPanGestureRecognizerHandler(sender) : oldTransitionPanGestureRecognizerHandler(sender)
-    }
-
-    func newTransitionPanGestureRecognizerHandler(_ sender: UIPanGestureRecognizer) {
-            guard let miniPlayer = appDelegate()?.miniPlayer(), !(miniPlayer.playerOpenState == .beingDragged || miniPlayer.playerOpenState == .animating) else { return }
-
-            if nowPlayingItem.timeSlider.isScrubbing() { return }
-
-            let touchPoint = sender.location(in: view?.window)
-
-            switch sender.state {
-            case .began:
-                initialTouchPoint = touchPoint
-            case .changed:
-                if touchPoint.y > initialTouchPoint.y {
-                    view.frame.origin.y = touchPoint.y - initialTouchPoint.y
-                }
-            case .ended, .cancelled:
-                // If pan ended, decide it we should close or reset the view
-                // based on the final position and the speed of the gesture
-                // (https://stackoverflow.com/a/47339617)
-                // If the swipe is too quick, we dismiss
-                let translation = sender.translation(in: view)
-                let velocity = sender.velocity(in: view)
-                let closing = (translation.y > view.frame.size.height * Self.minimumScreenRatioToHide) ||
-                (velocity.y > Self.minimumVelocityToHide) || velocity.y > 1000
-                dismissVelocity = velocity.y
-
-                if closing {
-                    finalYPositionWhenDismissing = touchPoint.y - initialTouchPoint.y
-                    miniPlayer.closeFullScreenPlayer()
-                } else {
-                    UIView.animate(withDuration: 0.2, animations: {
-                        self.view.frame = CGRect(x: 0,
-                                                 y: 0,
-                                                 width: self.view.frame.size.width,
-                                                 height: self.view.frame.size.height)
-                    })
-                }
-            default:
-                break
-            }
-        }
-
-    func oldTransitionPanGestureRecognizerHandler(_ sender: UIPanGestureRecognizer) {
         guard let miniPlayer = appDelegate()?.miniPlayer(), !(miniPlayer.playerOpenState == .beingDragged || miniPlayer.playerOpenState == .animating) else { return }
 
         if nowPlayingItem.timeSlider.isScrubbing() { return }
@@ -62,18 +17,30 @@ extension PlayerContainerViewController: UIGestureRecognizerDelegate {
         case .began:
             initialTouchPoint = touchPoint
         case .changed:
-            if touchPoint.y - initialTouchPoint.y > 0 {
-                let yPosition = touchPoint.y - initialTouchPoint.y
-                handleMoveTo(yPosition: yPosition, miniPlayer: miniPlayer)
+            if touchPoint.y > initialTouchPoint.y {
+                view.frame.origin.y = touchPoint.y - initialTouchPoint.y
             }
         case .ended, .cancelled:
-            if touchPoint.y - initialTouchPoint.y > PlayerContainerViewController.pullDownThreshold {
+            // If pan ended, decide it we should close or reset the view
+            // based on the final position and the speed of the gesture
+            // (https://stackoverflow.com/a/47339617)
+            // If the swipe is too quick, we dismiss
+            let translation = sender.translation(in: view)
+            let velocity = sender.velocity(in: view)
+            let closing = (translation.y > view.frame.size.height * Self.minimumScreenRatioToHide) ||
+            (velocity.y > Self.minimumVelocityToHide) || velocity.y > 1000
+            dismissVelocity = velocity.y
+
+            if closing {
+                finalYPositionWhenDismissing = touchPoint.y - initialTouchPoint.y
                 miniPlayer.closeFullScreenPlayer()
             } else {
-                UIView.animate(withDuration: Constants.Animation.defaultAnimationTime) {
-                    self.view.moveTo(y: 0)
-                    miniPlayer.moveToHiddenTopPosition()
-                }
+                UIView.animate(withDuration: 0.2, animations: {
+                    self.view.frame = CGRect(x: 0,
+                                             y: 0,
+                                             width: self.view.frame.size.width,
+                                             height: self.view.frame.size.height)
+                })
             }
         default:
             break
@@ -88,9 +55,7 @@ extension PlayerContainerViewController: UIGestureRecognizerDelegate {
     func handleScrollViewDidScroll(scrollView: UIScrollView) {
         guard let miniPlayer = appDelegate()?.miniPlayer(), !(miniPlayer.playerOpenState == .beingDragged || miniPlayer.playerOpenState == .animating) else { return }
 
-        if scrollView.contentOffset.y >= 0 {
-            miniPlayer.moveToHiddenTopPosition()
-        } else {
+        if scrollView.contentOffset.y < 0 {
             let yPosition = floor(-scrollView.contentOffset.y)
             handleMoveTo(yPosition: yPosition, miniPlayer: miniPlayer)
         }
@@ -105,7 +70,6 @@ extension PlayerContainerViewController: UIGestureRecognizerDelegate {
             let bottomSafeAreaOffset = window.safeAreaInsets.bottom
             let deviceSpecificPadding = bottomSafeAreaOffset > 0 ? (bottomSafeAreaOffset / 2) : -UIUtil.statusBarHeight(in: window)
             let offset = view.frame.minY - view.frame.size.height + miniPlayer.view.bounds.height + deviceSpecificPadding
-            miniPlayer.moveWhileDragging(offsetFromTop: offset)
         }
 
         if yPosition > PlayerContainerViewController.pullDownThreshold {